refactor: moved EVM types to pkg/evm
