+++ pkg/chains/evm/types_test.go
@@ -1,16 +1,13 @@
-package types_test
+package evm
 
 import (
 	"encoding/json"
 	"testing"
-
-	"github.com/shinzonetwork/shinzo-generator-client/pkg/chains/evm"
-	"github.com/shinzonetwork/shinzo-generator-client/pkg/types"
 )
 
 func TestTransactionReceiptJSONMarshaling(t *testing.T) {
 	t.Parallel()
-	receipt := types.TransactionReceipt{
+	receipt := TransactionReceipt{
 		TransactionHash:   "0x1234567890abcdef",
 		TransactionIndex:  "0",
 		BlockHash:         "0xabcdef1234567890",
@@ -21,23 +18,20 @@ func TestTransactionReceiptJSONMarshaling(t *testing.T) {
 		GasUsed:           "21000",
 		ContractAddress:   "0xcontract",
 		Status:            "0x1",
-		Logs:              []types.Log{},
+		Logs:              []Log{},
 	}
 
-	// Test marshaling
 	data, err := json.Marshal(receipt)
 	if err != nil {
 		t.Fatalf("Failed to marshal TransactionReceipt: %v", err)
 	}
 
-	// Test unmarshaling
-	var unmarshaled types.TransactionReceipt
+	var unmarshaled TransactionReceipt
 	err = json.Unmarshal(data, &unmarshaled)
 	if err != nil {
 		t.Fatalf("Failed to unmarshal TransactionReceipt: %v", err)
 	}
 
-	// Verify data integrity
 	if unmarshaled.TransactionHash != receipt.TransactionHash {
 		t.Errorf("TransactionHash mismatch: got %s, want %s", unmarshaled.TransactionHash, receipt.TransactionHash)
 	}
@@ -48,7 +42,7 @@ func TestTransactionReceiptJSONMarshaling(t *testing.T) {
 
 func TestBlockJSONMarshaling(t *testing.T) {
 	t.Parallel()
-	block := types.Block{
+	block := Block{
 		Hash:             "0x1234567890abcdef",
 		Number:           "12345",
 		Timestamp:        "1600000000",
@@ -64,23 +58,20 @@ func TestBlockJSONMarshaling(t *testing.T) {
 		TransactionsRoot: "0xtxroot",
 		ReceiptsRoot:     "0xreceiptsroot",
 		ExtraData:        "extra",
-		Transactions:     []types.Transaction{},
+		Transactions:     []Transaction{},
 	}
 
-	// Test marshaling
 	data, err := json.Marshal(block)
 	if err != nil {
 		t.Fatalf("Failed to marshal Block: %v", err)
 	}
 
-	// Test unmarshaling
-	var unmarshaled types.Block
+	var unmarshaled Block
 	err = json.Unmarshal(data, &unmarshaled)
 	if err != nil {
 		t.Fatalf("Failed to unmarshal Block: %v", err)
 	}
 
-	// Verify data integrity
 	if unmarshaled.Hash != block.Hash {
 		t.Errorf("Hash mismatch: got %s, want %s", unmarshaled.Hash, block.Hash)
 	}
@@ -91,7 +82,7 @@ func TestBlockJSONMarshaling(t *testing.T) {
 
 func TestTransactionJSONMarshaling(t *testing.T) {
 	t.Parallel()
-	tx := types.Transaction{
+	tx := Transaction{
 		Hash:             "0xtxhash",
 		BlockHash:        "0xblockhash",
 		BlockNumber:      "12345",
@@ -104,23 +95,20 @@ func TestTransactionJSONMarshaling(t *testing.T) {
 		Nonce:            "1",
 		TransactionIndex: 0,
 		Status:           true,
-		Logs:             []types.Log{},
+		Logs:             []Log{},
 	}
 
-	// Test marshaling
 	data, err := json.Marshal(tx)
 	if err != nil {
 		t.Fatalf("Failed to marshal Transaction: %v", err)
 	}
 
-	// Test unmarshaling
-	var unmarshaled types.Transaction
+	var unmarshaled Transaction
 	err = json.Unmarshal(data, &unmarshaled)
 	if err != nil {
 		t.Fatalf("Failed to unmarshal Transaction: %v", err)
 	}
 
-	// Verify data integrity
 	if unmarshaled.Hash != tx.Hash {
 		t.Errorf("Hash mismatch: got %s, want %s", unmarshaled.Hash, tx.Hash)
 	}
@@ -131,7 +119,7 @@ func TestTransactionJSONMarshaling(t *testing.T) {
 
 func TestLogJSONMarshaling(t *testing.T) {
 	t.Parallel()
-	log := types.Log{
+	log := Log{
 		Address:          "0xcontract",
 		Topics:           []string{"0xtopic1", "0xtopic2"},
 		Data:             "0xlogdata",
@@ -143,20 +131,17 @@ func TestLogJSONMarshaling(t *testing.T) {
 		Removed:          false,
 	}
 
-	// Test marshaling
 	data, err := json.Marshal(log)
 	if err != nil {
 		t.Fatalf("Failed to marshal Log: %v", err)
 	}
 
-	// Test unmarshaling
-	var unmarshaled types.Log
+	var unmarshaled Log
 	err = json.Unmarshal(data, &unmarshaled)
 	if err != nil {
 		t.Fatalf("Failed to unmarshal Log: %v", err)
 	}
 
-	// Verify data integrity
 	if unmarshaled.Address != log.Address {
 		t.Errorf("Address mismatch: got %s, want %s", unmarshaled.Address, log.Address)
 	}
@@ -165,116 +150,5 @@ func TestLogJSONMarshaling(t *testing.T) {
 	}
 	if unmarshaled.Removed != log.Removed {
 		t.Errorf("Removed mismatch: got %t, want %t", unmarshaled.Removed, log.Removed)
-	}
-}
-
-func TestRequestJSONMarshaling(t *testing.T) {
-	t.Parallel()
-	request := types.Request{
-		Type:  "query",
-		Query: "{ Block { number } }",
-	}
-
-	// Test marshaling
-	data, err := json.Marshal(request)
-	if err != nil {
-		t.Fatalf("Failed to marshal Request: %v", err)
-	}
-
-	// Test unmarshaling
-	var unmarshaled types.Request
-	err = json.Unmarshal(data, &unmarshaled)
-	if err != nil {
-		t.Fatalf("Failed to unmarshal Request: %v", err)
-	}
-
-	// Verify data integrity
-	if unmarshaled.Type != request.Type {
-		t.Errorf("Type mismatch: got %s, want %s", unmarshaled.Type, request.Type)
-	}
-	if unmarshaled.Query != request.Query {
-		t.Errorf("Query mismatch: got %s, want %s", unmarshaled.Query, request.Query)
-	}
-}
-
-func TestResponseJSONMarshaling(t *testing.T) {
-	t.Parallel()
-	response := types.Response{
-		Data: map[string][]struct {
-			DocID string `json:"_docID"`
-		}{
-			evm.CollectionBlock: {
-				{DocID: "doc-id-1"},
-				{DocID: "doc-id-2"},
-			},
-		},
-	}
-
-	// Test marshaling
-	data, err := json.Marshal(response)
-	if err != nil {
-		t.Fatalf("Failed to marshal Response: %v", err)
-	}
-
-	// Test unmarshaling
-	var unmarshaled types.Response
-	err = json.Unmarshal(data, &unmarshaled)
-	if err != nil {
-		t.Fatalf("Failed to unmarshal Response: %v", err)
-	}
-
-	// Verify data integrity
-	if len(unmarshaled.Data[evm.CollectionBlock]) != 2 {
-		t.Errorf("Expected 2 Block entries, got %d", len(unmarshaled.Data[evm.CollectionBlock]))
-	}
-	if unmarshaled.Data[evm.CollectionBlock][0].DocID != "doc-id-1" {
-		t.Errorf("DocID mismatch: got %s, want %s", unmarshaled.Data[evm.CollectionBlock][0].DocID, "doc-id-1")
-	}
-}
-
-func TestUpdateStructsJSONMarshaling(t *testing.T) {
-	t.Parallel()
-	// Test UpdateTransactionStruct
-	updateTx := types.UpdateTransactionStruct{
-		BlockId: "block-id-123",
-		TxHash:  "0xtxhash123",
-	}
-
-	data, err := json.Marshal(updateTx)
-	if err != nil {
-		t.Fatalf("Failed to marshal UpdateTransactionStruct: %v", err)
-	}
-
-	var unmarshaledTx types.UpdateTransactionStruct
-	err = json.Unmarshal(data, &unmarshaledTx)
-	if err != nil {
-		t.Fatalf("Failed to unmarshal UpdateTransactionStruct: %v", err)
-	}
-
-	if unmarshaledTx.BlockId != updateTx.BlockId {
-		t.Errorf("BlockId mismatch: got %s, want %s", unmarshaledTx.BlockId, updateTx.BlockId)
-	}
-
-	// Test UpdateLogStruct
-	updateLog := types.UpdateLogStruct{
-		BlockId:  "block-id-123",
-		TxId:     "tx-id-456",
-		TxHash:   "0xtxhash123",
-		LogIndex: "0",
-	}
-
-	data, err = json.Marshal(updateLog)
-	if err != nil {
-		t.Fatalf("Failed to marshal UpdateLogStruct: %v", err)
-	}
-
-	var unmarshaledLog types.UpdateLogStruct
-	err = json.Unmarshal(data, &unmarshaledLog)
-	if err != nil {
-		t.Fatalf("Failed to unmarshal UpdateLogStruct: %v", err)
-	}
-
-	if unmarshaledLog.TxId != updateLog.TxId {
-		t.Errorf("TxId mismatch: got %s, want %s", unmarshaledLog.TxId, updateLog.TxId)
 	}
 }
